jrobin 1.5.9.1 sha1sum 7d76f5cdec625fd1ce21acc0e46d286fbd6d307f Released 2008-09-05 Downloaded from http://sourceforge.net/projects/jrobin/files/jrobin/1.5.9/ License: LGPL 2.1
It appears that the only difference from 1.5.9 is it was compiled with Java 5
instead of Java 6. The 1.5.9 jars out there also contain the large font .ttf files,
which are not in 1.5.9.1.

The only announcement I could find about 1.5.9.1 is here:
http://opennms.530661.n2.nabble.com/Am-I-missing-something-td841945.html

The following directories are dropped from the jar during the build process:
org/jrobin/cmd/
org/jrobin/convertor/
org/jrobin/inspector/