Use bzip2 to compress debian packages
This mainly has an effect on the size of the javadocs package (i2p-doc shrinks
to half of its present size when using bzip2).  xz would give better
compression, of course, but xz support isn't isn't available in the version of
dpkg that ships with Lucid. All supported versions of Debian and Ubuntu support
bzip2 in .debs, so this will only have the impact of giving us smaller javadoc
packages.