Re-Write window.onbeforeunload
I would suggest removing this event entirely. Its usage is questionable amongst browsers, to say the least, and its affect to the end user is negligible and rough, if possibly even undesired sometimes. And since its not even used on the cancel button, which I would think would be the obvious choice, then it rarely gets executed.   If removal is chosen: 1) we can also remove the `onClick` code in the `button()` generator function, 2) delete `compose.js` entirely and the code that links it, 3) Remove the translation entries from the database, as well as 4) remove the event itself... But if you must keep it, here is my commit to fix it and read on:
----
`window.onbeforeunload` does not work as expected, nor intuitively. Check the MDN ( https://developer.mozilla.org/en-US/docs/Web/Events/beforeunload ).  The MDN basically says that no modern browser supports using the string that is supposed to be returned. The return value is also ignored on Edge in my testing. Funny thing, the MDN has an "almost cross-browser" solution, but then goes on to say that it doesn't matter in most modern browsers. And some further thinking, most other browsers not mentioned are usually forks from those mentioned. Making it virtually impossible to find a browser that supports actually using the contents of that return value. I couldn't find one in my testing.

Furthermore, it is highly suggested to use `addEventListener` for `beforeunload` (and.. well.. really.. for all events). So, with this commit I switched to using `addEventListener`. And instead of `removeEventListener` on the buttons. I opted to keep the event out there and just toggle its functionality.  This actually fixes a crash with the `button()` generator function. For instance, the cancel button is also used on the Setting screen of Susimail. But there is no function declaration here for `cancelPopup()` since it was not linked in. So it crashes. Now, any button can run its new `onClick` payload of `beforePopup = false;", without error,  because it will just create that variable if it doesn't already exist.  This was actually where I started. I was chasing down this crash.
----
A few notes:

-`compose.js` in susimail is no longer needed. Its contents were only one line anyhow. There is no reason to link one line, one that is also not reused on another page. And since the functionality was not working as intended; I just got rid of it in my re-write. It really is a lot simpler/safer/faster/smaller to just set a boolean.

-TODO: The text that was getting translated should be removed from the Message database. It was not getting displayed anyhow. And I've not looked into how  to clean up that database.

-`beforeunload` does not look at the contents of `e.returnValue` nor `return <value>` very deeply. Basically, if you set `e.returnValue` to anything, then it will show the popup. Otherwise, if you leave it alone, it will not. I assume this behavior is mimicked by the regular `return <value>`... but I have no clue.. it does not seem required.

-This commit also fixes some situations in Edge.

-If you click `New` for a new email. And then don't type anything, and then immediately click the back button. It will attempt to generate this popup prompt. In Chrome, it will suppress the prompt if the user has not first made any meaningful gestures on the page, whilst generating an error about it in the console. However, in Edge, and some other browsers, it will not suppress and it will show the popup regardless of user interaction.

PT