Still WIP, but I've added a new sbt plugin named sbt-native-packager for future rpm/deb/windows and maybe even OSX packages. Also, the macosx sbt project now has a task named buildAppBundleTask and which will produce an I2P.app and copy over needed directories/jars/wars which later needs to be extracted to an writeable area which would be i2p base directory in runtime.
The MacOSXRouterLauncherApp contains more information about how 
the executable in the OS X bundle will locate our R/O i2pbase so 
it can copy it to a writable area. The R/O is also to ensure valid
signature on the bundle.

At last, this approach is done casue letting an Mach-O binary load
libjvm.dylib was proved unstable, so MacOSXRouterLauncherApp will 
rather use an JNI module to load the needed glue with the 
Mac OS X system.