Change json lib from net.minidev.json to com.json.simple because libjson-simple-java is in Debian (jessie+) and Ubuntu (trusty+). Version 1.1.1 (Feb. 2012) from: https://github.com/cliftonlabs/json-simple https://github.com/fangyidong/json-simple/releases License: Apache 2 (same as minidev) Release info: https://cliftonlabs.github.io/json-simple/ Current version 2.3.0 (in buster/sid/bionic+) is compatible. The 3.x branch is incompatible.
No known external users of minidev (JSON_SMART) which was added in April 2018
(0.9.35) for DoH and the eventual bundling of i2pcontrol.
The i2pcontrol plugin was switched from json-simple to minidev in early 2018,
but it bundles the minidev source. When we bundle, we can switch it back.

Does not compile; see checkins to follow.