Debian: Create jammy build to deal with Jetty API change (Gitlab #496)
Jetty made a breaking API change to PathSpec in 9.4.30, so
focal (with 9.4.26) is not compatible with jammy+ and buster+ (9.4.45+).

Symptom is failure to load translated help pages in console.

Fix is to make focal only for focal; create jammy build
for jammy+ and buster+. Update build docs.

Does not affect non-package builds, where we bundle Jetty 9.3.x.