Startup: Add migration tool to migrate all Jetty xml files
referenced in clients.config command lines.
This changes Ref id to Ref refid, and updates the DTD.
Previous changes for this were for new installs only,
this handles existing files in the eepsite dirs.

This builds on the existing MigrateJetty class,
while removing the ancient Jetty 5-8 to 9 migration code.

The only way that would have been used these days is
if somebody tried to update directly from 0.9.29 or earlier
(8 years ago) to the current release.
With this change, we tell the user to uninstall and reinstall.