Router: Fix PublishRouterInfoJob sometimes not getting started
if setNetDbReady() is called when netDb().isInitialized() is false.
This can happen if the router has a few RIs at at startup but
not enough and it reseeds, or if reseed is very slow?
If PLRIJ does not run then the local RI doesn't get updated
and most connections fail due to RI clock skew in the handshake.
Then the router is essentially stuck.